fix(viewer): light the Cesium GLB so walls/reveals don't look mis-cut
The Cesium 3D-map overlay built its merged GLB with only POSITION + COLOR_0 and
a KHR_materials_unlit material. Unlit + no normals means Cesium can't shade the
building, so correct recessed faces (window/door reveals, opening cuts) render
as flat strips that look like wrong cuts in the outer walls. The geometry is
NOT wrong — buildMergedGLB reuses the exact same canonical Rust MeshData the
WebGPU viewer renders (void cuts baked in); only the GLB representation dropped
shading.

- Emit a NORMAL attribute (per-vertex, direction — not origin-shifted; +Z
  fallback for any mesh lacking normals).
- Replace the unlit material with a lit, double-sided PBR material
  (baseColorFactor white so the COLOR_0 vertex colours show, lit by Cesium's
  sun). Double-sided since IFC winding isn't reliably outward.
- Extract buildMergedGLB out of the CesiumOverlay component into a pure,
  testable module (apps/viewer/src/lib/geo/cesium-glb.ts).

Tests: parse the GLB container and assert NORMAL present, lit PBR material
(not unlit), self-consistent accessor/bufferView byte math, origin folded into
world positions. typecheck clean.

Note: geometry export is already canonical Rust — viewer, GLTFExporter and this
Cesium GLB all consume the same process_geometry MeshData; only this packer
needed normals + lighting, no geometry-pipeline change.
